reconcile "old" and "new" S2I config properties
This commit removes some pieces from the "old" S2I config that
were added recently (by mistake). It also reverts the default
value of `s2i.jar-directory` in the "old" S2I config, because
the `container-image-s2i` extension still builds the image
so that application files are in `/deployments`.

In addition, this commit also fixes the way how `S2IConfig`
(the "new" S2I config) and `OpenshiftConfig` are merged.